sdk: provide ABI-dependent atree files
Backport from AOSP master

This patch allows development/sdk/ to provide ABI-specific .atree
files. This is useful to copy prebuilt CPU model-specific files
to the SDK platform image directory.

This is needed to deal with the fact that for ARMv7 platform builds,
the kernel image to use is prebuilt/android-arm/kernel-qemu-armv7 instead
of prebuilt/android-arm/kernel/qemu

Change-Id: Ib76caa54a059eca2dafc9cdba812b29f98b5e3c3
Orig-Change-Id: Icdca751852c01421fccf196436a9bc1114ea90f7